% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationalMatrix.R
\name{Qdet}
\alias{Qdet}
\title{Determinant of a rational matrix}
\usage{
Qdet(M)
}
\arguments{
\item{M}{a square matrix such that \code{as.character(Mij)} is 
a quoted integer or a quoted fraction for each entry \code{Mij}}
}
\value{
A string: quoted rational number representing the determinant.
}
\description{
Determinant of a square matrix with rational entries.
}
\examples{
library(RationalMatrix)
M <- cbind(c("1/2", "3"), c("5/3", "-2/7"))
Qdet(M)
}
