% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationalMatrix.R
\name{Qinverse}
\alias{Qinverse}
\title{Inverse of a rational matrix}
\usage{
Qinverse(M)
}
\arguments{
\item{M}{a square matrix such that \code{as.character(Mij)} is 
a quoted integer or a quoted fraction for each entry \code{Mij}}
}
\value{
A character matrix representing the inverse of \code{M}.
}
\description{
Inverse matrix of a square rational matrix.
}
\examples{
library(RationalMatrix)
M <- cbind(c("1/2", "3", "1"), c("5/3", "-2/7", "10/3"), c("0", "1", "2"))
Qinverse(M)
}
