% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rationalMatrix.R
\name{Qrank}
\alias{Qrank}
\title{Rank of a rational matrix}
\usage{
Qrank(M)
}
\arguments{
\item{M}{a matrix such that \code{as.character(Mij)} is 
a quoted integer or a quoted fraction for each entry \code{Mij}}
}
\value{
An integer, the rank of \code{M}.
}
\description{
Returns the rank of a rational matrix.
}
\examples{
library(RationalMatrix)
M <- cbind(c("1/2", "3", "1"), c("5/3", "-2/7", "10/3"), c("1", "1", "2"))
Qrank(M)
}
